requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.35.2
torch==2.2.2
numpy==1.26.4
python-dotenv==1.0.0
haversine==2.8.0
pydantic==2.5.0
streamlit==1.28.0
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
audio_recorder_streamlit==0.0.6
datasets==2.14.6
accelerate==0.24.1
sentencepiece==0.2.0
protobuf==3.20.3
openai-whisper==20231117 
geopy==2.4.1
folium==0.14.0
python-json-logger==2.0.7